Include required ClientID header for twitch
Twitch made Client-ID a required header for each request so that
they can throttle bad clients. Hope that stream center isn't
such a bad client.

Decided to implement this by NOT checking in the value to git because
I do not want Twitch to get angry at me for leaking my ClientID