Update for swift 3
Wow there was a lot of required changes. Most notably, fixing a lot of
string munging and range building code that feels more fragile than not.

Also learned that you have to be REAL CAREFUL when adding constraints with
xcode 8 -- you get a runtime error if you don't take care to explicitly say
that your constraints dictionary is a [String: UIVIew]

There is also a fair amount of remaining effort

* getting tests to compile again
* getting tests to pass again
* Verify that Chat, Pause, Menu, Search features still work
* Remove all of the "optional"s that crept into the UI during this upgrade
* Audit and remove some of the extensions on builtins (like dictionary, string, etc)
* Auditing usage of AlamoFire and other libraries -- can we do better?